Use delta time logic. Misc refactoring.
Probably going to stick with this implementation for a game loop. It
should be fast enough to not worry about de-coupling the game logic from
the render code. The trade-off, I think, is simplicity for decreased
performance (more CPU intensive, more battery use, etc).

Alternative would be to try and clamp the re-render intervals, or
de-couple the loop to calculate game logic from each render.