More work on collisions.
This factors in sign when detecting collision. This makes it easy to not
care if objects overlap over multiple simulation frame. The first flips
the sign and the second no longer detects a collision.

Another quick workaround is that the sign is y velocity sign is flipped
when the ball hits the top segment of a paddle. This means There is no
top edge detection. If it hits the top edge it also hits the side edge
and will be high enough up the paddle to change directions.

Depending on the collision distance from the ball to the center of the
paddle the y velocity is modifed. The function attempts to increase the
max speed in y as the collision distance from the center of the ball to
the center of the paddle increases. In theory a corner hit should set
the velocity to max. The one improvement here is that the y velocity
feels like it "levels" out too much with close to center hits.