Clean up collision code.
This just abstracts everything to make it re-usable and readable. Two
improvements can be made here. First this can all be namespaced. There
is still some internal knowledge in some of the functions so didn't want
to just pull everything out yet. Second, when more things start
happening on collision "events" this code will break and the abstraction
will break down. It is probably fine for the scope of this project but a
better seperation of the collision event and the effects would be needed
for anything larger.